refactor(render): Rename src to lib.
affects: tessellate-render
